Fix infinite loop when executing 'lan print' command
The retrycount counter is currently not incremented, so in the situation
when this command fails, library enters infinite loop of executing this
command, which is bad.

This commit adds missing counter incrementation with the same style as
in basecommands and make it return current @info in case of a faliure,
so all methods accessing it will just get return nil.